requirements.txt
# Agentic Security Lab - Python Dependencies
# ============================================
# Requires Python 3.9 or higher

# Core Dependencies
python-dotenv>=0.19.0,<2.0.0
requests>=2.31.0
pydantic>=2.5.0

# Vault Integration
hvac>=2.1.0

# JWT Handling
PyJWT>=2.8.0
cryptography>=41.0.0

# Database
psycopg2-binary>=2.9.0
SQLAlchemy>=2.0.0

# Langflow (compatible with Python 3.9)
# Note: Langflow 1.x requires Python 3.10+, using 0.6.x for broader compatibility
langflow>=0.6.0,<0.7.0

# Jupyter Notebook
jupyter>=1.0.0
ipykernel>=6.27.0
ipywidgets>=8.1.0

# Data Visualization
matplotlib>=3.8.0
pandas>=2.0.0

# Logging and Monitoring
colorlog>=6.8.0

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0

# Development Tools (optional)
black>=23.12.0
flake8>=7.0.0
mypy>=1.7.0